Improving API readability

+++ julez-core/src/main/java/uk/co/acuminous/julez/scenario/source/BaseScenarioSource.java
@@ -6,7 +6,7 @@ import uk.co.acuminous.julez.scenario.ScenarioSource;
 
 public abstract class BaseScenarioSource implements ScenarioSource {
 
-    public SizeLimiter limitRepetitionsTo(int repetitions) {
+    public SizeLimiter limitRepetitionsTo(long repetitions) {
         return new SizeLimiter(this, repetitions);
     }
